+++ src/main/java/chess/domain/chessboard/Numbering.java
@@ -1,4 +1,4 @@
-package chess.chessboard;
+package chess.domain.chessboard;
 
 public enum Numbering {
