+++ src/main/java/chess/domain/chessboard/Square.java
@@ -1,4 +1,4 @@
-package chess.chessboard;
+package chess.domain.chessboard;
 
 import java.util.Objects;
