+++ src/main/java/chess/domain/chesspiece/Camp.java
@@ -1,4 +1,4 @@
-package chess.chesspiece;
+package chess.domain.chesspiece;
 
 public enum Camp {
