+++ src/main/java/chess/domain/chesspiece/movestrategy/KingMoveStrategy.java
@@ -1,4 +1,4 @@
-package chess.chesspiece.movestrategy;
+package chess.domain.chesspiece.movestrategy;
 
 public class KingMoveStrategy implements MoveStrategy{
