+++ src/main/java/chess/domain/chesspiece/movestrategy/MoveStrategy.java
@@ -1,4 +1,4 @@
-package chess.chesspiece.movestrategy;
+package chess.domain.chesspiece.movestrategy;
 
 public interface MoveStrategy {
