+++ src/main/java/chess/domain/chesspiece/movestrategy/PawnMoveStrategy.java
@@ -1,4 +1,4 @@
-package chess.chesspiece.movestrategy;
+package chess.domain.chesspiece.movestrategy;
 
 public class PawnMoveStrategy implements MoveStrategy {
